+++ packages/webcomponents/src/components/payouts-list/payouts-list-core.tsx
@@ -5,9 +5,19 @@ import { TableEmptyState, TableErrorState, TableLoadingState } from '../../ui-co
 import { payoutTableCells, payoutTableColumns } from './payouts-table';
 import { Table } from '../../utils/table';
 import { getRequestParams, onQueryParamsChange } from './payouts-list-params-state';
-import { table, tableCell } from '../../styles/parts';
 import { ComponentClickEvent, ComponentErrorEvent } from '../../api/ComponentEvents';
-import { TableClickActions } from '../../ui-components/table/event-types';
+import {
+  TableWrapper,
+  TableComponent,
+  TableHead,
+  TableHeadRow,
+  TableBody,
+  TableRow,
+  TableFoot,
+  TableFootRow,
+  TableFootCell,
+  TableClickActions
+} from '../../ui-components';
 
 @Component({
   tag: 'payouts-list-core',
@@ -172,14 +182,14 @@ export class PayoutsListCore {
 
   render() {
     return (
-      <div class="table-wrapper" part="table-wrapper">
-        <table class="table table-hover" part={table}>
-          <thead class="table-head sticky-top" part="table-head">
-            <tr class="table-light text-nowrap" part="table-head-row">
+      <TableWrapper>
+        <TableComponent>
+          <TableHead>
+            <TableHeadRow>
               {this.payoutsTable.columnData.map((column) => column)}
-            </tr>
-          </thead>
-          <tbody class="table-body">
+            </TableHeadRow>
+          </TableHead>
+          <TableBody>
             <TableLoadingState
               columnSpan={this.payoutsTable.columnKeys.length}
               isLoading={this.loading}
@@ -193,19 +203,19 @@ export class PayoutsListCore {
               errorMessage={this.errorMessage}
             />
             {this.showRowData && this.payoutsTable.rowData.map((data, index) => (
-              <tr
+              <TableRow
                 data-test-id="table-row"
                 data-row-entity-id={this.entityId[index]}
                 onClick={this.rowClickHandler}
               >
                 {data}
-              </tr>
+              </TableRow>
             ))}
-          </tbody>
+          </TableBody>
           {this.paging && (
-            <tfoot class="sticky-bottom">
-              <tr class="table-light align-middle">
-                <td part={tableCell} colSpan={this.payoutsTable.columnData.length}>
+            <TableFoot>
+              <TableFootRow>
+                <TableFootCell colSpan={this.payoutsTable.columnData.length}>
                   <pagination-menu
                     paging={{
                       ...this.paging,
@@ -213,12 +223,12 @@ export class PayoutsListCore {
                       handleClickNext: this.handleClickNext,
                     }}
                   />
-                </td>
-              </tr>
-            </tfoot>
+                </TableFootCell>
+              </TableFootRow>
+            </TableFoot>
           )}
-        </table>
-      </div>
+        </TableComponent>
+      </TableWrapper>
     );
   }
 }