+++ packages/webcomponents/src/ui-components/index.ts
@@ -1,10 +1,6 @@
 export * from './button';
 export * from './button-spinner';
 export * from './styled-host/styled-host';
-export * from './table/exported-parts';
-export * from './table-empty-state';
-export * from './table-error-state';
-export * from './table-loading-state';
 export * from './skeleton';
 export * from './headers/header-1';
 export * from './headers/header-2';
@@ -12,3 +8,4 @@ export * from './headers/header-3';
 export * from './form/form-helpers/form-control-help-text';
 export * from './form/form-helpers/form-control-error-text';
 export * from './details/details';
+export * from './table';